Adding yals.h updated

+++ yals.h
@@ -30,33 +30,33 @@ Yals * yals_new_with_mem_mgr (void*, YalsMalloc, YalsRealloc, YalsFree);
 /*------------------------------------------------------------------------*/
 
 void yals_srand (Yals *, unsigned long long seed);
-int yals_setopt (Yals *, const char * name, int val);
+long long yals_setopt (Yals *, const char * name, long long val);
 void yals_setprefix (Yals *, const char *);
 void yals_setout (Yals *, FILE *);
-void yals_setphase (Yals *, int lit);
+void yals_setphase (Yals *, long long lit);
 void yals_setflipslimit (Yals *, long long);
 void yals_setmemslimit (Yals *, long long);
 
-int yals_getopt (Yals *, const char * name);
+long long yals_getopt (Yals *, const char * name);
 void yals_usage (Yals *);
 void yals_showopts (Yals *);
 
 /*------------------------------------------------------------------------*/
 
-void yals_add (Yals *, int lit);
+void yals_add (Yals *, long long lit);
 
-int yals_sat (Yals *);
+long long yals_sat (Yals *);
 
 /*------------------------------------------------------------------------*/
 
 long long yals_flips (Yals *);
 long long yals_mems (Yals *);
 
-int yals_minimum (Yals *);
-int yals_lkhd (Yals *);
-int yals_deref (Yals *, int lit);
+long long yals_minimum (Yals *);
+long long yals_lkhd (Yals *);
+long long yals_deref (Yals *, long long lit);
 
-const int * yals_minlits (Yals *);
+const long long * yals_minlits (Yals *);
 
 /*------------------------------------------------------------------------*/
 
@@ -64,7 +64,7 @@ void yals_stats (Yals *);
 
 /*------------------------------------------------------------------------*/
 
-void yals_seterm (Yals *, int (*term)(void*), void*);
+void yals_seterm (Yals *, long long (*term)(void*), void*);
 
 void yals_setime (Yals *, double (*time)(void));
